+++ spirit/user/auth/tests/tests.py
@@ -12,7 +12,7 @@ from django.test.utils import override_settings
 from django.core.urlresolvers import NoReverseMatch
 
 from ....core.tests import utils
-from ..forms import RegistrationForm, ResendActivationForm
+from ..forms import RegistrationForm, ResendActivationForm, LoginForm
 from ..backends import EmailAuthBackend
 from ...utils.tokens import UserActivationTokenGenerator
 from ...models import UserProfile
@@ -142,8 +142,12 @@ class UserViewTest(TestCase):
         """
         test access
         """
-        response = self.client.get(reverse('spirit:user:auth:password-reset-confirm', kwargs={'uidb64': 'f-a-k-e',
-                                                                                    'token': 'f-a-k-e'}))
+        response = self.client.get(
+            reverse(
+                'spirit:user:auth:password-reset-confirm',
+                kwargs={'uidb64': 'f-a-k-e', 'token': 'f-a-k-e'}
+            )
+        )
         self.assertEqual(response.status_code, 200)
 
     def test_admin_login(self):
@@ -171,8 +175,12 @@ class UserViewTest(TestCase):
         self.user.is_active = False
         self.user.save()
         token = UserActivationTokenGenerator().generate(self.user)
-        response = self.client.get(reverse('spirit:user:auth:registration-activation', kwargs={'pk': self.user.pk,
-                                                                                     'token': token}))
+        response = self.client.get(
+            reverse(
+                'spirit:user:auth:registration-activation',
+                kwargs={'pk': self.user.pk, 'token': token}
+            )
+        )
         expected_url = reverse("spirit:user:auth:login")
         self.assertRedirects(response, expected_url, status_code=302)
         self.assertTrue(User.objects.get(pk=self.user.pk).is_active)
@@ -188,8 +196,12 @@ class UserViewTest(TestCase):
         utils.login(self)
         User.objects.filter(pk=self.user.pk).update(is_active=False)
         UserProfile.objects.filter(user__pk=self.user.pk).update(is_verified=True)
-        response = self.client.get(reverse('spirit:user:auth:registration-activation', kwargs={'pk': self.user.pk,
-                                                                                     'token': token}))
+        response = self.client.get(
+            reverse(
+                'spirit:user:auth:registration-activation',
+                kwargs={'pk': self.user.pk, 'token': token}
+            )
+        )
         expected_url = reverse("spirit:user:auth:login")
         self.assertRedirects(response, expected_url, status_code=302)
         self.assertFalse(User.objects.get(pk=self.user.pk).is_active)
@@ -302,12 +314,28 @@ class UserFormTest(TestCase):
         form = RegistrationForm(data=form_data)
         self.assertEqual(form.is_valid(), True)
 
+    def test_registration_login(self):
+        """
+        Register and login
+        """
+        form_data = {'username': 'foo', 'email': 'foo@foo.com',
+                     'password1': 'pass123', 'password2': 'pass123'}
+        form = RegistrationForm(data=form_data)
+        self.assertEqual(form.is_valid(), True)
+
+        user = form.save()
+        self.assertFalse(user.is_active)
+
+        user.is_active = True
+        user.save()
+        utils.login(self, user=user, password='pass123')  # Asserts if can't login
+
     def test_registration_email_required(self):
         """
         Registration should require the email field
         """
         form_data = {'username': 'foo',
-                     'password1': 'pass', 'password2': 'pass'}
+                     'password': 'pass'}
         form = RegistrationForm(data=form_data)
         self.assertEqual(form.is_valid(), False)
         self.assertIn('email', form.errors)
@@ -356,6 +384,37 @@ class UserFormTest(TestCase):
                      'password1': 'pass123', 'password2': 'pass123'}
         form = RegistrationForm(data=form_data)
         self.assertEqual(form.is_valid(), True)
+
+    def test_registration_email_confirmation(self):
+        """
+        Confirmation email should match email
+        """
+        form_data = {'username': 'foo', 'email': 'foo@bar.com',
+                     'email2': 'foofoo@bar.com', 'password': 'pass'}
+        form = RegistrationForm(data=form_data)
+        self.assertEqual(form.is_valid(), False)
+        self.assertNotIn('email2', form.cleaned_data)
+
+    @override_settings(ST_CASE_INSENSITIVE_EMAILS=True)
+    def test_registration_email_confirmation_case_insensitive(self):
+        """
+        Confirmation email should match email
+        """
+        form_data = {'username': 'foo', 'email': 'FOO@bar.com',
+                     'password1': 'pass123', 'password2': 'pass123'}
+        form = RegistrationForm(data=form_data)
+        self.assertEqual(form.is_valid(), True)
+
+    @override_settings(ST_CASE_INSENSITIVE_EMAILS=False)
+    def test_registration_email_confirmation_case_sensitive(self):
+        """
+        Confirmation email should match email
+        """
+        form_data = {'username': 'foo', 'email': 'FOO@bar.com',
+                     'email2': 'FOO@BAR.COM', 'password': 'pass'}
+        form = RegistrationForm(data=form_data)
+        self.assertEqual(form.is_valid(), False)
+        self.assertNotIn('email2', form.cleaned_data)
 
     def test_resend_activation_email(self):
         """
@@ -414,6 +473,69 @@ class UserFormTest(TestCase):
         form = ResendActivationForm(form_data)
         self.assertFalse(form.is_valid())
         self.assertRaises(AttributeError, form.get_user)
+
+    def test_login(self):
+        """
+        Should login the user
+        """
+        utils.create_user(username="foobar", password="foo")
+        form_data = {'username': "foobar", 'password': "foo"}
+        form = LoginForm(data=form_data)
+        self.assertTrue(form.is_valid())
+
+    def test_login_email(self):
+        """
+        Should login the user by email
+        """
+        utils.create_user(email="foobar@bar.com", password="foo")
+        form_data = {'username': "foobar@bar.com", 'password': "foo"}
+        form = LoginForm(data=form_data)
+        self.assertTrue(form.is_valid())
+
+    @override_settings(ST_CASE_INSENSITIVE_EMAILS=False)
+    def test_login_email_case_sensitive(self):
+        """
+        Should login the user by email
+        """
+        utils.create_user(email="foobar@bar.com", password="foo")
+        form_data = {'username': "FOOBAR@bar.com", 'password': "foo"}
+        form = LoginForm(data=form_data)
+        self.assertFalse(form.is_valid())
+
+    @override_settings(ST_CASE_INSENSITIVE_EMAILS=True)
+    def test_login_email_case_sensitive(self):
+        """
+        Should login the user by email
+        """
+        utils.create_user(email="foobar@bar.com", password="foo")
+        form_data = {'username': "FOOBAR@bar.com", 'password': "foo"}
+        form = LoginForm(data=form_data)
+        self.assertTrue(form.is_valid())
+
+    def test_login_invalid(self):
+        """
+        Should not login invalid user
+        """
+        form = LoginForm(data={})
+        self.assertFalse(form.is_valid())
+
+    def test_login_password_invalid(self):
+        """
+        Should not login invalid user
+        """
+        utils.create_user(username="foobar", password="foo")
+        form_data = {'username': "foobar", 'password': "bad"}
+        form = LoginForm(data=form_data)
+        self.assertFalse(form.is_valid())
+
+    def test_login_username_invalid(self):
+        """
+        Should not login invalid user
+        """
+        utils.create_user(username="foobar", password="foo")
+        form_data = {'username': "bad", 'password': "foo"}
+        form = LoginForm(data=form_data)
+        self.assertFalse(form.is_valid())
 
 
 class UserBackendTest(TestCase):